reorganize components
